refactor: extract peer event and data channel logic into usePeerDataHandlers hook
